- Fixed a few 32/64 type issues in OWF.

+++ Frameworks/OWF/Content.subproj/Address.subproj/OWAddress.m
@@ -717,7 +717,7 @@ addressForNotSoObviousHostname(NSString *string)
 
 // Equality and hashing
 
-- (unsigned)hash;
+- (NSUInteger)hash;
 {
     return [url hash];
 }
